requirements.txt
pandas==1.4.3
polars==0.19.3
numpy==1.21.6
scikit-learn==1.3.2
matplotlib==3.5.1
catboost==1.2.3
pytorch-lightning==2.0.2
transformers==4.44.2
sentence_transformers==3.1.1
torchmetrics==0.11.3
torchinfo==1.7.2
pytorch-lifestream